third_party/WebKit/LayoutTests/virtual/mediasession/media/mediasession/htmlmediaelement-set-session-expected.txt

CONSOLE WARNING: line 9: dummy log entry to avoid an empty -expected.txt in virtual/
This is a testharness.js-based test.
FAIL HTMLMediaElement set session Failed to construct 'MediaSession': Missing platform implementation.
Harness: the test ran to completion.
